# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_thepantheon_api_internal_model.CreateVadeMecumRequest:
    properties:
      cabecalho:
        type: string
      capitulo:
        type: string
      category:
        type: string
      content:
        type: string
      description:
        type: string
      file_url:
        type: string
      idcapitulo:
        type: string
      idtitulo:
        type: string
      textocapitulo:
        type: string
      textodotitulo:
        type: string
      title:
        minLength: 3
        type: string
      titulo:
        type: string
    required:
    - cabecalho
    - capitulo
    - category
    - description
    - idcapitulo
    - idtitulo
    - textocapitulo
    - textodotitulo
    - title
    - titulo
    type: object
...
  github_com_thepantheon_api_internal_model.UpdateVadeMecumRequest:
    properties:
      cabecalho:
        type: string
      capitulo:
        type: string
      category:
        type: string
      content:
        type: string
      description:
        type: string
      file_url:
        type: string
      idcapitulo:
        type: string
      idtitulo:
        type: string
      textocapitulo:
        type: string
      textodotitulo:
        type: string
      title:
        minLength: 3
        type: string
      titulo:
        type: string
    type: object
...
  github_com_thepantheon_api_internal_model.VadeMecum:
    properties:
      cabecalho:
        type: string
      capitulo:
        type: string
      category:
        type: string
      content:
        type: string
      created_at:
        type: string
      description:
        type: string
      file_url:
        type: string
      id:
        type: string
      idcapitulo:
        type: string
      idtitulo:
        type: string
      textocapitulo:
        type: string
      textodotitulo:
        type: string
      title:
        type: string
      titulo:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pantheonconcursos.com.br
    name: API Support
  description: API REST para sistema de concursos com autenticação JWT
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: ThePantheon API
  version: "1.0"
paths:
...
  /vade-mecum:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecum'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar itens de Vade Mecum
      tags:
      - vade-mecum
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do Vade Mecum
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.CreateVadeMecumRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecum'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar item de Vade Mecum
      tags:
      - vade-mecum
  /vade-mecum/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remover Vade Mecum
      tags:
      - vade-mecum
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecum'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obter Vade Mecum por ID
      tags:
      - vade-mecum
    put:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Dados para atualização
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.UpdateVadeMecumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecum'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar Vade Mecum
      tags:
      - vade-mecum
  /vade-mecum/category/{category}:
    get:
      parameters:
      - description: Categoria
        in: path
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecum'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar itens por categoria
      tags:
      - vade-mecum
    post:
      consumes:
      - application/json
      parameters:
      - description: Categoria
        in: path
        name: category
        required: true
        type: string
      - description: Dados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.CreateVadeMecumRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecum'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar item específico por categoria
      tags:
      - vade-mecum
  /vade-mecum/category/{category}/{id}:
    delete:
      parameters:
      - description: Categoria
        in: path
        name: category
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remover item de uma categoria
      tags:
      - vade-mecum
    put:
      consumes:
      - application/json
      parameters:
      - description: Categoria
        in: path
        name: category
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Dados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_thepantheon_api_internal_model.UpdateVadeMecumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_thepantheon_api_internal_model.VadeMecum'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar item dentro de uma categoria
      tags:
      - vade-mecum
...
securityDefinitions:
  Bearer:
    description: Digite "Bearer" seguido do token JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"